# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CommonDto:
    properties:
      message:
        type: string
    type: object
host: 127.0.0.1:8081
info:
  contact:
    email: 1014007642@qq.com
    name: yzy
    url: https://www.yucoding.club
  description: 基于区块链的煤炭大数据物流平台的后端接口文档
  title: 2508yyds队后端接口文档
  version: "1.0"
paths:
  /auth/info:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommonDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CommonDto'
      summary: 获取个人信息
      tags:
      - 用户管理
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 手机号
        in: query
        name: telephone
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommonDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CommonDto'
      summary: 用户登录
      tags:
      - 用户管理
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户名
        in: query
        name: name
        required: true
        type: string
      - description: 电话号码
        in: query
        name: telephone
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 权限级别 2.煤炭场 3.普通用户
        in: query
        name: kind
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommonDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CommonDto'
      summary: 用户注册
      tags:
      - 用户管理
...
  /chain/invoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 方法名
        in: query
        name: name
        required: true
        type: string
      - description: 参数json
        in: query
        name: args
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommonDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CommonDto'
      summary: 调用链码
      tags:
      - 链码相关
  /chain/query:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 方法名
        in: query
        name: name
        required: true
        type: string
      - description: 参数列表
        in: query
        items:
          type: string
        name: args
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommonDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.CommonDto'
      summary: 调用链码
      tags:
      - 链码相关
...
swagger: "2.0"